Stefan Fiedler:         update kde (3.5.5), kdevelop (3.3.5)- rediffed 	remove kdelibs version hotfix
[2006111616421211322] (https://www.rocklinux.net/submaster)



git-svn-id: http://www.rocklinux.org/svn/rock-linux/trunk@8018 c5f82cb5-29bc-0310-9cd0-bff59a50e3bc

+++ package/kde/update-kde.sh
@@ -5,13 +5,16 @@ if [ ! "$1" -o ! "$2" ] ; then
 	exit -1
 fi
 
-for x in kde* ; do
+for x in kde* arts ; do
+	[ -f $x/$x.desc ] || continue
 	echo "Updating $x ..."
 	sed 	-e s,$1,$2,g \
 		-e "s/\[D\] [0-9]* /\[D\] 0 /" $x/$x.desc > $x/$x.desc.new
 	mv $x/$x.desc.new $x/$x.desc
 done
 
+# The version of arts is 1.x.y, but the download dir contains 3.x.y, hence
+# the arts.desc is changed above and below.
 x="arts"
 echo "Updating $x ..."
 sed 	-e s,1.${1:2},1.${2:2},g \
